change bigint to int

+++ src/main/scala/com/rhdzmota/fbmessenger/webhook/model/Entry.scala
@@ -2,4 +2,4 @@ package com.rhdzmota.fbmessenger.webhook.model
 
 import com.rhdzmota.fbmessenger.webhook.model.message.Message
 
-case class Entry(id: String, time: BigInt, messaging: List[Message])
+case class Entry(id: String, time: Int, messaging: List[Message])
