+++ src/main/scala/com/rhdzmota/fbmessenger/webhook/model/message/Message.scala
@@ -1,3 +1,3 @@
 package com.rhdzmota.fbmessenger.webhook.model.message
 
-case class Message(sender: Participant, recipient: Participant, timestamp: BigInt, message: MessageData)
+case class Message(sender: Participant, recipient: Participant, timestamp: Int, message: MessageData)
